requirements.txt
Pillow
numpy
cairosvg